requirements.txt
streamlit-drawable-canvas-fix
Pillow
numpy
